backend/requirements.txt
django>=4.2.0,<4.3.0
djangorestframework
psycopg2